Rework tox config
This is done because there is no way to make tox to install something in
virtualenv before it starts to install deps section. And it basically
means that you HAVE TO HAVE access to some PyPI. Now tox installs only
setup requirement and other stuff is installed by pip.

This is life.