Fix problems with playbook_dir
ceph-ansible actively uses playbook_dir variable and in some cases it
prevents Ansible from finding of correct file (even though it was
introduced to solve this problem :((( ). But Ansible 2.2.1.0 cannot
correctly set playbook_dir to required path if role is connected with
role_path in config (the path will be set to the playbook.yaml, this is
reasonable, but we do not expected that at all).

The solution is to have a symlink on correct location. We cannot
redefine playbook_dir therefore there is no many choices.