Support for NIC aliases on generating playbooks
In Ansible, there is a mess on interface facts: some of them have
device, activeness and interface, some are not. This patchset adds
tricky logic on management of NIC aliases but this is not robust because
it uses some euristics on determining if interface is alias (there is
no way in Ansible 2.2 to know if interface is alias of a certain
interface).

Change-Id: If623f7f6bcdd8a628cc9f5dca2225796d512c5bb