Fix for cluster redeployment
This is a huge commit which includes a bunch of changes related to
cluster redeployment. This is done as a single commit since it almost
makes no sense to rework it into several: changes are tightly related to
each other.

The thing is: with latest Jewel packages we've got slightly changed
behaviour which affects frequent cluster redeployments only. To support
all cases we had to update to newer ceph-ansible which has a relevant
changes.

But we have to introduce a list of fixes/hacks around:

1. We have to disable StartLimitBurst in systemd on install because
   monitors can restart a number of times waiting for the keys. It is
   ok, because we reenable bursting after deployment.

2. ceph-ansible did a various changes for key generation, therefore
   we have to make an attempt to regenerate keys if they are absent
   explicitly. Version v2.2.0 of ceph-ansible has a bug when
   they are using with_items instead of with_flattened.

3. 0.pool.ntp.org caused a list of issues, switched on pool.ntp.org
   for proper load balancing

4. update to latest <2.3 Ansible (this helps to defeat a list of flaws
   with support of v2.2.0 of ceph-ansible)

5. ceph-ansible starts to verify how cluster works after deployment.
   This has helped to eliminate a list of small bugs.

Change-Id: I1f8cdafc4b22dd007b0e107b2e5a6c89a58beb51